Изменил и дописал логику заказов и заказов для админа, добавил транзации, дто для добавления и удаления заказов, добавил использование findOrFail, добавил его в обработчик исключений и др. исправления

+++ database/migrations/2024_06_02_160207_create_orders_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
                 $table->increments('id');
                 $table->string('status');
                 $table->string('address');
-                $table->float('total');
+                $table->integer('total');
                 $table->string('phone');
                 $table->string('title');
                 $table->foreignId('user_id');
